requirements.txt
pillow
requests
openpyxl
pyinstaller
xlsxwriter
